frontend/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-auth==2.35.0
python-dotenv==1.0.1
httpx==0.27.2
